+++ src/core/device_stomp.c
@@ -300,23 +300,23 @@ int DEVICE_STOMP_StartAllConnections(void)
 **
 ** Function called to queue a message on the specified STOMP connection
 **
-** \param   usp_msg_type - Type of USP message contained in pbuf. This is used for debug logging when the message is sent by the MTP.
+** \param   msi - Information about the content to send. The ownership of
+**                the payload buffer is passed to this function, unless an error is returned.
 ** \param   instance - instance number of the stomp connection in Device.STOMP.Connection.{i}
 ** \param   controller_queue - name of STOMP queue to send this message to
 ** \param   agent_queue - name of agent's STOMP queue configured for this connection in the data model.
 **                        NOTE: This may be NULL, if agent's STOMP queue is set by subscribe_dest: STOMP header
-** \param   pbuf - pointer to buffer containing binary protobuf message. Ownership of this buffer passes to this code, if successful
-** \param   pbuf_len - length of buffer containing protobuf binary message
 ** \param   err_id_header - pointer to string containing the STOMP usp-err-id header
 ** \param   expiry_time - time at which the USP message should be removed from the MTP send queue
 **
 ** \return  USP_ERR_OK if successful
 **
 **************************************************************************/
-int DEVICE_STOMP_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int instance, char *controller_queue, char *agent_queue, unsigned char *pbuf, int pbuf_len, char *err_id_header, time_t expiry_time)
+int DEVICE_STOMP_QueueBinaryMessage(mtp_send_item_t *msi, int instance, char *controller_queue, char *agent_queue, char *err_id_header, time_t expiry_time)
 {
-    int err;
+    int err = USP_ERR_GENERAL_FAILURE;
     stomp_conn_params_t *sp;
+    USP_ASSERT(msi != NULL);
 
     // Exit if unable to find the specified STOMP connection
     sp = FindStompParamsByInstance(instance);
@@ -327,7 +327,7 @@ int DEVICE_STOMP_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int insta
     }
 
     // Exit if unable to queue the message
-    err = STOMP_QueueBinaryMessage(usp_msg_type, instance, controller_queue, agent_queue, pbuf, pbuf_len, kMtpContentType_UspRecord, err_id_header, expiry_time);
+    err = STOMP_QueueBinaryMessage(msi, instance, controller_queue, agent_queue, err_id_header, expiry_time);
     if (err != USP_ERR_OK)
     {
         return err;
@@ -915,6 +915,9 @@ int NotifyChange_StompUsername(dm_req_t *req, char *value)
 {
     stomp_conn_params_t *sp;
     bool schedule_reconnect = false;
+    char buf[MAX_DM_SHORT_VALUE_LEN];
+    dm_vendor_get_mtp_username_cb_t   get_mtp_username_cb;
+    int err;
 
     // Determine stomp connection to be updated
     sp = FindStompParamsByInstance(inst1);
@@ -924,6 +927,23 @@ int NotifyChange_StompUsername(dm_req_t *req, char *value)
     if ((strcmp(sp->username, value) != 0) && (sp->enable))
     {
         schedule_reconnect = true;
+    }
+
+    // Override a blank username in the database with that provided by a core vendor hook
+    if (*value == '\0')
+    {
+        get_mtp_username_cb = vendor_hook_callbacks.get_mtp_username_cb;
+        if (get_mtp_username_cb != NULL)
+        {
+            // Exit if vendor hook failed
+            err = get_mtp_username_cb(inst1, buf, sizeof(buf));
+            if (err != USP_ERR_OK)
+            {
+                return err;
+            }
+
+            value = buf;
+        }
     }
 
     // Set the new value. This must be done before scheduling a reconnect, so that the reconnect uses the correct values
@@ -1254,6 +1274,9 @@ int ProcessStompConnAdded(int instance)
     stomp_conn_params_t *sp;
     int err;
     char path[MAX_DM_PATH];
+    char buf[MAX_DM_SHORT_VALUE_LEN];
+    dm_vendor_get_mtp_username_cb_t   get_mtp_username_cb;
+    dm_vendor_get_mtp_password_cb_t   get_mtp_password_cb;
 
     // Exit if unable to add another STOMP connection
     sp = FindUnusedStompParams();
@@ -1298,6 +1321,25 @@ int ProcessStompConnAdded(int instance)
         goto exit;
     }
 
+    // Override a blank username in the database with that provided by a core vendor hook
+    if (sp->username[0] == '\0')
+    {
+        get_mtp_username_cb = vendor_hook_callbacks.get_mtp_username_cb;
+        if (get_mtp_username_cb != NULL)
+        {
+            // Exit if vendor hook failed
+            err = get_mtp_username_cb(instance, buf, sizeof(buf));
+            if (err != USP_ERR_OK)
+            {
+                goto exit;
+            }
+
+            // Replace the blank password from the database with the password retrieved via core vendor hook
+            USP_SAFE_FREE(sp->username);
+            sp->username = USP_STRDUP(buf);
+        }
+    }
+
     // Exit if unable to get the password for this STOMP connection
     USP_SNPRINTF(path, sizeof(path), "%s.%d.Password", device_stomp_conn_root, instance);
     err = DM_ACCESS_GetPassword(path, &sp->password);
@@ -1309,9 +1351,6 @@ int ProcessStompConnAdded(int instance)
     // Override a blank password in the database with that provided by a core vendor hook
     if (sp->password[0] == '\0')
     {
-        char buf[MAX_DM_SHORT_VALUE_LEN];
-        dm_vendor_get_mtp_password_cb_t   get_mtp_password_cb;
-
         get_mtp_password_cb = vendor_hook_callbacks.get_mtp_password_cb;
         if (get_mtp_password_cb != NULL)
         {