+++ src/core/mqtt.c
@@ -1,8 +1,8 @@
 /*
  *
- * Copyright (C) 2019-2021, Broadband Forum
+ * Copyright (C) 2019-2022, Broadband Forum
  * Copyright (C) 2020, BT PLC
- * Copyright (C) 2020-2021  CommScope, Inc
+ * Copyright (C) 2020-2022  CommScope, Inc
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -108,7 +108,7 @@ typedef struct
     double_linked_list_t usp_record_send_queue;
 
     // From the broker
-    mqtt_subscription_t response_subscription;
+    mqtt_subscription_t response_subscription; // NOTE: The topic in here may be an empty string if not set by either Device.LocalAgent.MTP.{i}.ResponseTopicConfigured or present in the CONNACK
 
     int retry_count;
     time_t retry_time;
@@ -120,6 +120,7 @@ typedef struct
     // Scheduler
     mqtt_conn_params_t next_params;
     scheduled_action_t schedule_reconnect;   // Sets whether an MQTT reconnect is scheduled
+    scheduled_action_t schedule_close;       // Sets whether an MQTT disable is scheduled
 
     STACK_OF(X509) *cert_chain; // Certificate chain saved during SSL cert verification, and used to determine the role for the controller
     ssl_verify_callback_t *verify_callback;
@@ -128,9 +129,7 @@ typedef struct
     bool are_certs_loaded;      // Flag indicating whether the above ssl_ctx has been loaded with the trust store certs
                                 // It is used to ensure that the certs are loaded only once, rather than on every reconnect
 
-    bool in_tcp_connect;        // Set when calling the mosquitto_connect (and MQTT v5 equivalent).
-                                // This flag causes the data model thread to delay performing any disconnect until after the connect call has returned
-    bool disconnect_after_tcp_connect; // Set if the MQTT thread should perform a disconnect after returning from the MQTT connect call
+    char *agent_topic_from_connack;  // Saved copy of agent's topic (if received in the CONNACK)
 
 } mqtt_client_t;
 
@@ -146,12 +145,12 @@ char *mqtt_state_names[kMqttState_Max] =
     "Error/Retring"
 };
 
+//------------------------------------------------------------------------------
+// Payload to send in MQTT queue
 typedef struct
 {
     double_link_t link;     // Doubly linked list pointers. These must always be first in this structure
-    Usp__Header__MsgType usp_msg_type;  // Type of USP message contained within pbuf
-    unsigned char *pbuf;    // Protobuf format message to send in binary format
-    int pbuf_len;           // Length of protobuf message to send
+    mtp_send_item_t item;   // Information about the content to send
     char *topic;            // Name of the MQTT Topic to send to
     mqtt_qos_t qos;         // QOS to request when sending message
     int mid;                // MQTT message ID
@@ -171,6 +170,7 @@ void MoveState_Private(mqtt_state_t *state, mqtt_state_t to, const char *event, 
 void HandleMqttError(mqtt_client_t *client, mqtt_failure_t failure_code, const char* message);
 int DisableMqttClient(mqtt_client_t *client, bool is_reconnect);
 void FreeMqttClientCertChain(mqtt_client_t *client);
+void SaveAgentTopicFromConnack(mqtt_client_t *client, char *agent_topic);
 
 //------------------------------------------------------------------------------------
 // Callbacks
@@ -459,6 +459,7 @@ int PerformMqttClientConnect(mqtt_client_t *client)
     mosquitto_property *proplist = NULL;
     int mosq_err = MOSQ_ERR_SUCCESS;
     int err = USP_ERR_OK;
+    int keep_alive;
 
     // Exit if unable to configure username/password for this mosquitto context
     if (strlen(client->conn_params.username) > 0)
@@ -499,10 +500,19 @@ int PerformMqttClientConnect(mqtt_client_t *client)
         }
     }
 
+    // Calculate the keep alive period to pass to libmosquitto. We might need to alter this, because libmosquitto does not support keep alive < 5 seconds
+    keep_alive = client->conn_params.keepalive;
+    if (keep_alive == 0)
+    {
+        keep_alive = 60*60*18;  // Set to 18 hours which is the largest that libmosquitto accepts (it truncates the arg to uint16 internally)
+    }
+    else if (keep_alive < 5)
+    {
+        keep_alive = 5;
+    }
+
     // Release the access mutex temporarily whilst performing the connect call
     // We do this to prevent the data model thread from potentially being blocked, whilst the connect call is taking place
-    client->in_tcp_connect = true;
-    USP_ASSERT(client->disconnect_after_tcp_connect == false);
     OS_UTILS_UnlockMutex(&mqtt_access_mutex);
 
     // Perform the TCP connect
@@ -510,26 +520,16 @@ int PerformMqttClientConnect(mqtt_client_t *client)
     if (version == kMqttProtocol_5_0)
     {
         mosq_err = mosquitto_connect_bind_v5(client->mosq, client->conn_params.host, client->conn_params.port,
-                                             client->conn_params.keepalive, NULL, proplist);
+                                             keep_alive, NULL, proplist);
     }
     else
     {
         mosq_err = mosquitto_connect(client->mosq, client->conn_params.host, client->conn_params.port,
-                                     client->conn_params.keepalive);
+                                     keep_alive);
     }
 
     // Take the access mutex again
     OS_UTILS_LockMutex(&mqtt_access_mutex);
-    client->in_tcp_connect = false;
-
-    // If the MQTT client was scheduled to be disabled by another thread, whilst we were in the connect call, then perform the disable here
-    if (client->disconnect_after_tcp_connect)
-    {
-        client->disconnect_after_tcp_connect = false;
-        DisableMqttClient(client, false);       // NOTE: Deliberately ignoring any errors. Probably because client already in disconnected state because above connect call failed
-        err = USP_ERR_OK;
-        goto exit;
-    }
 
     // Exit if failed to connect
     if (mosq_err != MOSQ_ERR_SUCCESS)
@@ -567,12 +567,6 @@ void Connect(mqtt_client_t *client)
 
     // Start the MQTT Connect
     err = PerformMqttClientConnect(client);
-
-    // Exit if the MQTT client was disabled by another thread during the TCP connect
-    if (client->state == kMqttState_Idle)
-    {
-        return;
-    }
 
     // Exit if failed to connect
     if (err != USP_ERR_OK)
@@ -718,10 +712,18 @@ int Unsubscribe(mqtt_client_t *client, mqtt_subscription_t *sub)
     return err;
 }
 
-int SubscribeToAll(mqtt_client_t *client)
+void SubscribeToAll(mqtt_client_t *client)
 {
-    int err = USP_ERR_OK;
     int i;
+    char buf[128];
+
+    // Exit if no agent response topic configured (or set by the CONNACK)
+    if ((client->response_subscription.topic==NULL) || (client->response_subscription.topic[0] == '\0'))
+    {
+        USP_SNPRINTF(buf, sizeof(buf), "%s: No response topic configured (or set by the CONNACK)", __FUNCTION__);
+        HandleMqttError(client, kMqttFailure_Misconfigured, buf);
+        return;
+    }
 
     // Let the DM know we're ready for sending messages
     DM_EXEC_PostMqttHandshakeComplete(client->conn_params.instance, client->role);
@@ -740,7 +742,8 @@ int SubscribeToAll(mqtt_client_t *client)
 
             if (Subscribe(client, sub) != USP_ERR_OK)
             {
-                err = USP_ERR_INTERNAL_ERROR;
+                USP_SNPRINTF(buf, sizeof(buf), "%s: mosquitto_subscribe() failed for topic=%s", __FUNCTION__, sub->topic);
+                HandleMqttError(client, kMqttFailure_OtherError, buf);
             }
         }
     }
@@ -748,11 +751,9 @@ int SubscribeToAll(mqtt_client_t *client)
     // Subscribe to response topic too
     if (Subscribe(client, &client->response_subscription) != USP_ERR_OK)
     {
-        USP_LOG_Error("%s: Failed to subscribe to response topic", __FUNCTION__);
-        err = USP_ERR_INTERNAL_ERROR;
+        USP_SNPRINTF(buf, sizeof(buf), "%s: mosquitto_subscribe() failed for agent's response topic=%s", __FUNCTION__, client->response_subscription.topic);
+        HandleMqttError(client, kMqttFailure_OtherError, buf);
     }
-
-    return err;
 }
 
 int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
@@ -768,6 +769,7 @@ int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
         goto error;
     }
 
+    USP_ASSERT((client->response_subscription.topic!= NULL) && (client->response_subscription.topic[0] != '\0')); // SubscribeToAll() should have prevented the code getting here
     if (mosquitto_property_add_string(&proplist, RESPONSE_TOPIC, client->response_subscription.topic) != MOSQ_ERR_SUCCESS)
     {
         USP_LOG_Error("%s: Failed to add response topic string", __FUNCTION__);
@@ -783,7 +785,7 @@ int PublishV5(mqtt_client_t *client, mqtt_send_item_t *msg)
         goto error;
     }
 
-    int mosq_err = mosquitto_publish_v5(client->mosq, &msg->mid, msg->topic, msg->pbuf_len, msg->pbuf, msg->qos, false /* retain */, proplist);
+    int mosq_err = mosquitto_publish_v5(client->mosq, &msg->mid, msg->topic, msg->item.pbuf_len, msg->item.pbuf, msg->qos, false /* retain */, proplist);
     if (mosq_err != MOSQ_ERR_SUCCESS)
     {
         USP_LOG_Error("%s: Failed to publish to v5 with error %d", __FUNCTION__, mosq_err);
@@ -805,7 +807,7 @@ int Publish(mqtt_client_t *client, mqtt_send_item_t *msg)
     USP_ASSERT(msg != NULL);
     USP_ASSERT(msg->topic != NULL);
 
-    MSG_HANDLER_LogMessageToSend(msg->usp_msg_type, msg->pbuf, msg->pbuf_len, kMtpProtocol_MQTT, client->conn_params.host, NULL, kMtpContentType_UspRecord);
+    MSG_HANDLER_LogMessageToSend(&msg->item, kMtpProtocol_MQTT, client->conn_params.host, NULL);
 
     int version = client->conn_params.version;
     if (version == kMqttProtocol_5_0)
@@ -814,7 +816,7 @@ int Publish(mqtt_client_t *client, mqtt_send_item_t *msg)
     }
     else
     {
-        if (mosquitto_publish(client->mosq, &msg->mid, msg->topic, msg->pbuf_len, msg->pbuf, msg->qos, false /*retain*/) != MOSQ_ERR_SUCCESS)
+        if (mosquitto_publish(client->mosq, &msg->mid, msg->topic, msg->item.pbuf_len, msg->item.pbuf, msg->qos, false /*retain*/) != MOSQ_ERR_SUCCESS)
         {
             USP_LOG_Error("%s: Failed to publish to v3.1.1. Params:\n MID:%d\n topic:%s\n msg->qos:%d\n", __FUNCTION__, msg->mid, msg->topic, msg->qos);
             err = USP_ERR_INTERNAL_ERROR;
@@ -839,6 +841,7 @@ int DisconnectClient(mqtt_client_t *client)
         result = mosquitto_disconnect(client->mosq);
         if (result != MOSQ_ERR_SUCCESS)
         {
+            USP_LOG_Warning("%s: mosquitto_disconnect() returned error=%d", __FUNCTION__, result);
             err = USP_ERR_INTERNAL_ERROR;
         }
     }
@@ -912,7 +915,7 @@ void PopClientUspQueue(mqtt_client_t *client)
         if (head != NULL)
         {
             USP_SAFE_FREE(head->topic);
-            USP_SAFE_FREE(head->pbuf);
+            USP_SAFE_FREE(head->item.pbuf);
             DLLIST_Unlink(&client->usp_record_send_queue, head);
             USP_SAFE_FREE(head);
         }
@@ -972,7 +975,7 @@ bool IsUspRecordInMqttQueue(mqtt_client_t *client, unsigned char *pbuf, int pbuf
     q_msg = (mqtt_send_item_t *) client->usp_record_send_queue.head;
     while (q_msg != NULL)
     {
-        if ((q_msg->pbuf_len == pbuf_len) && (memcmp(q_msg->pbuf, pbuf, pbuf_len)==0))
+        if ((q_msg->item.pbuf_len == pbuf_len) && (memcmp(q_msg->item.pbuf, pbuf, pbuf_len)==0))
         {
             return true;
         }
@@ -1190,23 +1193,26 @@ void MessageV5Callback(struct mosquitto *mosq, void *userdata, const struct mosq
     }
     else
     {
-        USP_LOG_Info("%s: Received Message: Topic: %s", __FUNCTION__, message->topic);
-        //USP_LOG_Info("%s: Received Message: Payload: %s", __FUNCTION__, (char*)message->payload);
+        USP_LOG_Info("%s: Received Message: Topic: '%s' PayloadLength: %d bytes", __FUNCTION__, 
+                     message->topic, message->payloadlen);
 
         if (client->state == kMqttState_Running)
         {
             // Now we have a message from somewhere
-            char response_info[512] = { 0 };
-            char *response_info_ptr = response_info;
+            char *response_info_ptr = NULL;
 
             if (mosquitto_property_read_string(props, RESPONSE_TOPIC,
                     &response_info_ptr, false) == NULL)
             {
-                USP_LOG_Debug("%s: Failed to read response topic in message info: \"%s\"\n", __FUNCTION__, response_info_ptr);
-                response_info_ptr = NULL;
+                USP_LOG_Debug("%s: No controller response topic present in received message", __FUNCTION__);
             }
 
             ReceiveMqttMessage(client, message, response_info_ptr);
+
+            if (response_info_ptr != NULL)
+            {
+                free(response_info_ptr);
+            }
         }
         else
         {
@@ -1300,13 +1306,13 @@ void ConnectV5Callback(struct mosquitto *mosq, void *userdata, int result, int f
             free(client_id_ptr);
         }
 
+        // Update the agent topic (if received in this CONNACK)
+        USP_SAFE_FREE(client->agent_topic_from_connack);
         if (mosquitto_property_read_string(props, RESPONSE_INFORMATION,
               &response_info_ptr, false) != NULL)
         {
             // Then replace the response_topic in subscription with this
-            USP_SAFE_FREE(client->response_subscription.topic);
-            USP_LOG_Debug("%s: Received response_info: \"%s\"", __FUNCTION__, response_info_ptr);
-            client->response_subscription.topic = USP_STRDUP(response_info_ptr);
+            SaveAgentTopicFromConnack(client, response_info_ptr);
             free(response_info_ptr);
         }
         else
@@ -1319,9 +1325,7 @@ void ConnectV5Callback(struct mosquitto *mosq, void *userdata, int result, int f
                 // we only want subscribe-topic user property
                 if (strcmp("subscribe-topic", userPropName) == 0)
                 {
-                    USP_LOG_Debug("%s: Received subcribe-topic: \"%s\"", __FUNCTION__, subscribe_topic_ptr);
-                    USP_SAFE_FREE(client->response_subscription.topic);
-                    client->response_subscription.topic = USP_STRDUP(subscribe_topic_ptr);
+                    SaveAgentTopicFromConnack(client, subscribe_topic_ptr);
                     free(subscribe_topic_ptr);
                     free(userPropName);
                 }
@@ -1336,9 +1340,7 @@ void ConnectV5Callback(struct mosquitto *mosq, void *userdata, int result, int f
                         // we only want subscribe-topic user property
                         if (strcmp("subscribe-topic", userPropName) == 0)
                         {
-                            USP_LOG_Debug("%s: Received subcribe-topic: \"%s\"", __FUNCTION__, subscribe_topic_ptr);
-                            USP_SAFE_FREE(client->response_subscription.topic);
-                            client->response_subscription.topic = USP_STRDUP(subscribe_topic_ptr);
+                            SaveAgentTopicFromConnack(client, subscribe_topic_ptr);
                         }
                         free(subscribe_topic_ptr);
                         free(userPropName);
@@ -1346,6 +1348,7 @@ void ConnectV5Callback(struct mosquitto *mosq, void *userdata, int result, int f
                 }
             }
         }
+
         USP_LOG_Debug("%s: Received client id \"%s\"", __FUNCTION__, client->conn_params.client_id);
 
         ResetRetryCount(client);
@@ -1356,6 +1359,32 @@ void ConnectV5Callback(struct mosquitto *mosq, void *userdata, int result, int f
 
 exit:
     OS_UTILS_UnlockMutex(&mqtt_access_mutex);
+}
+
+/*********************************************************************//**
+**
+** SaveAgentTopicFromConnack
+**
+** Saves the agent topic received in the CONNACK into the MQTT client structure
+**
+** \param   client - pointer to MQTT client structure to update the agent topic in
+** \param   agent_topic - value of agent response topic received in the CONNACK
+**
+** \return  None
+**
+**************************************************************************/
+void SaveAgentTopicFromConnack(mqtt_client_t *client, char *agent_topic)
+{
+    USP_LOG_Debug("%s: Received agent-topic: \"%s\"", __FUNCTION__, agent_topic);
+
+    // Override agent response topic configured in Device.LocalAgent.MTP.{i}.MQTT.ResponseTopicConfigured
+    USP_SAFE_FREE(client->response_subscription.topic);
+    client->response_subscription.topic = USP_STRDUP(agent_topic);
+
+    // Save the agent response topic received in the CONNACK into the MQTT client structure
+    // (so it can be read by Device.LocalAgent.MTP.{i}.MQTT.ResponseTopicDiscovered and Device.MQTT.Client.{i}.ResponseInformation)
+    USP_SAFE_FREE(client->agent_topic_from_connack);
+    client->agent_topic_from_connack = USP_STRDUP(agent_topic);
 }
 
 //------------------------------------------------------------------------------
@@ -1628,7 +1657,8 @@ void MessageCallback(struct mosquitto *mosq, void *userdata, const struct mosqui
     }
     else
     {
-        USP_LOG_Info("%s: Received Message: Topic: %s Payload: %s", __FUNCTION__, message->topic, (char*)message->payload);
+        USP_LOG_Info("%s: Received Message: Topic: '%s' PayloadLength: %d bytes", __FUNCTION__, 
+                     message->topic, message->payloadlen);
 
         if (client->state == kMqttState_Running)
         {
@@ -1682,7 +1712,7 @@ void DisconnectCallback(struct mosquitto *mosq, void *userdata, int rc)
     {
         if (client->state != kMqttState_ErrorRetrying)
         {
-            // We have successfully, gracefully disconnected from the broker
+            // We have successfully performed an agent-initiated disconnect from the broker
             MoveState(&client->state, kMqttState_Idle, "Disconnected from broker - ok");
         }
         else
@@ -1757,11 +1787,13 @@ void InitClient(mqtt_client_t *client, int index)
     client->mosq = NULL;
     client->role = ROLE_DEFAULT;
     client->schedule_reconnect = kScheduledAction_Off;
+    client->schedule_close = kScheduledAction_Off;
     client->cert_chain = NULL;
     client->verify_callback = mqtt_verify_callbacks[index];
     client->socket_fd = INVALID;
     client->ssl_ctx = NULL;   // NOTE: The SSL context is created in MQTT_Start()
     client->are_certs_loaded = false;
+    client->agent_topic_from_connack = NULL;
 
     ResetRetryCount(client);
 
@@ -1933,10 +1965,11 @@ void MQTT_Stop(void)
 }
 
 // Called when you already have the mqtt access mutex and a valid client
-int EnableClient(mqtt_client_t* client)
+int EnableMqttClient(mqtt_client_t* client)
 {
     USP_ASSERT(client != NULL);
     client->schedule_reconnect = kScheduledAction_Off;
+    client->schedule_close = kScheduledAction_Off;
 
     // Add response topic as a new "subscription"
     mqtt_subscription_t resp_sub = { 0 };
@@ -2011,7 +2044,7 @@ int MQTT_EnableClient(mqtt_conn_params_t *mqtt_params, mqtt_subscription_t subsc
 
     if (client->conn_params.enable)
     {
-        err = EnableClient(client);
+        err = EnableMqttClient(client);
     }
 
 exit:
@@ -2089,7 +2122,10 @@ int DisableMqttClient(mqtt_client_t *client, bool is_reconnect)
     }
 
     // Tell libmosquitto to disconnect from the broker
-    err = DisconnectClient(client);
+    if (client->state != kMqttState_ErrorRetrying)
+    {
+        err = DisconnectClient(client);
+    }
 
     // Free all member variables (unless they're needed for a reconnect)
     CleanMqttClient(client, is_reconnect);
@@ -2129,8 +2165,8 @@ int MQTT_DisableClient(int instance, bool is_reconnect)
         goto exit;
     }
 
-    // Exit if the client is already in the Idle state after a graceful disconnect, freeing the rest of the structure
-    // (which we couldn't do at the time of the graceful disconnect, because we needed the instance number to persist)
+    // Exit if the client is already in the Idle state after an agent-initiated disconnect, freeing the rest of the structure
+    // (which we couldn't do at the time of the disconnect, because we needed the instance number to persist)
     if (client->state == kMqttState_Idle)
     {
         CleanMqttClient(client, is_reconnect);
@@ -2138,18 +2174,9 @@ int MQTT_DisableClient(int instance, bool is_reconnect)
         goto exit;
     }
 
-    // Exit if the MQTT MTP thread is performing a blocking connect because libwebsockets won't allow the disconnect to happen at the same time
-    // Instead set a flag to perform the disable, after the blocking connect has returned
-    if (client->in_tcp_connect)
-    {
-        USP_ASSERT(is_reconnect == false);
-        client->disconnect_after_tcp_connect = true;
-        err = USP_ERR_OK;
-        goto exit;
-    }
-
-    // Disable the MQTT Client
-    err = DisableMqttClient(client, is_reconnect);
+    // Schedule the disable to occur after it has been activated, and all queued messages sent out
+    client->schedule_close = kScheduledAction_Signalled;
+    err = USP_ERR_OK;
 
 exit:
     OS_UTILS_UnlockMutex(&mqtt_access_mutex);
@@ -2163,10 +2190,10 @@ exit:
     return err;
 }
 
-int MQTT_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int instance, char* topic,
-        unsigned char *pbuf, int pbuf_len)
+int MQTT_QueueBinaryMessage(mtp_send_item_t *msi, int instance, char* topic)
 {
     int err = USP_ERR_GENERAL_FAILURE;
+    USP_ASSERT(msi != NULL);
 
     // Add the message to the back of the queue
     OS_UTILS_LockMutex(&mqtt_access_mutex);
@@ -2174,7 +2201,7 @@ int MQTT_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int instance, cha
     if (is_mqtt_mtp_thread_exited)
     {
         OS_UTILS_UnlockMutex(&mqtt_access_mutex);
-        USP_FREE(pbuf);
+        USP_FREE(msi->pbuf);
         return USP_ERR_OK;
     }
 
@@ -2190,21 +2217,17 @@ int MQTT_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int instance, cha
 
     // Find if this is a duplicate in the queue
     // May have been tried to be resent by the MTP_EXEC thread
-    if (IsUspRecordInMqttQueue(client, pbuf, pbuf_len))
+    if (IsUspRecordInMqttQueue(client, msi->pbuf, msi->pbuf_len))
     {
         // No error, just return success
-        USP_FREE(pbuf);
+        USP_FREE(msi->pbuf);
         err = USP_ERR_OK;
         goto exit;
     }
 
     mqtt_send_item_t *send_item;
     send_item = USP_MALLOC(sizeof(mqtt_send_item_t));
-    send_item->usp_msg_type = usp_msg_type;
-
-    // pbuf is our responsibility in MTP layer now
-    send_item->pbuf = pbuf;
-    send_item->pbuf_len = pbuf_len;
+    send_item->item = *msi;  // NOTE: Ownership of the payload buffer passes to the MQTT client
 
     if (topic != NULL)
     {
@@ -2213,6 +2236,18 @@ int MQTT_QueueBinaryMessage(Usp__Header__MsgType usp_msg_type, int instance, cha
     else
     {
         send_item->topic = USP_STRDUP(client->conn_params.topic);
+    }
+
+    // Exit (discarding the USP record) if no controller topic to send the message to
+    // NOTE: This should already have been ensured by the caller (in the function CalcNotifyDest)
+    if ((send_item->topic == NULL) || (send_item->topic[0] == '\0'))
+    {
+        USP_LOG_Error("%s: Discarding USP Message (%s) as no controller topic to send to", __FUNCTION__, MSG_HANDLER_UspMsgTypeToString(send_item->item.usp_msg_type));
+        USP_SAFE_FREE(send_item->item.pbuf);
+        USP_SAFE_FREE(send_item->topic);
+        USP_FREE(send_item);
+        err = USP_ERR_INTERNAL_ERROR;
+        goto exit;
     }
 
     send_item->mid = INVALID;
@@ -2284,6 +2319,33 @@ void MQTT_ProcessAllSocketActivity(socket_set_t* set)
                     }
                     break;
             }
+
+            // Deal with closing or restarting the connection (if all responses have been sent)
+            if (client->usp_record_send_queue.head == NULL)
+            {
+                if (client->schedule_reconnect == kScheduledAction_Activated)
+                {
+                    USP_LOG_Debug("%s: Schedule reconnect ready!", __FUNCTION__);
+
+                    // Stop the current client
+                    // NOTE: Intentionally ignoring any error returned from libmosquitto, since we cannot handle it
+                    DisableMqttClient(client, true);
+
+                    // Copy in the next_params, so that we have the correct conn_params for the next connection
+                    ParamReplace(&client->conn_params, &client->next_params);
+
+                    // Start a connection - through the normal C API
+                    EnableMqttClient(client);
+                }
+                else if (client->schedule_close == kScheduledAction_Activated)
+                {
+                    USP_LOG_Debug("%s: Schedule close ready!", __FUNCTION__);
+
+                    // Stop the current client
+                    // NOTE: Intentionally ignoring any error returned from libmosquitto, since we cannot handle it
+                    DisableMqttClient(client, false);
+                }
+            }
         }
     }
 
@@ -2329,21 +2391,6 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
                             USP_LOG_Error("%s: Failed to send head of the queue, leaving there to try again", __FUNCTION__);
                         }
                     }
-                    else if (client->schedule_reconnect == kScheduledAction_Activated)
-                    {
-                        // Responses would be sent if here
-                        USP_LOG_Debug("%s: Schedule reconnect ready!", __FUNCTION__);
-
-                        // Stop the current client
-                        MQTT_DisableClient(client->next_params.instance, true);
-
-                        // Copy in the next_params, so that we have the correct
-                        // conn_params for the next connection
-                        ParamReplace(&client->conn_params, &client->next_params);
-
-                        // Start a connection - through the normal C API
-                        EnableClient(client);
-                    }
                     break;
                 case kMqttState_ErrorRetrying:
                     {
@@ -2369,7 +2416,7 @@ void MQTT_UpdateAllSockSet(socket_set_t *set)
                         else if (client->retry_time - cur_time <= 0)
                         {
                             USP_LOG_Debug("%s: Retrying connection", __FUNCTION__);
-                            EnableClient(client);
+                            EnableMqttClient(client);
                         }
                         else
                         {
@@ -2441,6 +2488,8 @@ void MQTT_ActivateScheduledActions(void)
 {
     int i;
     mqtt_client_t* client;
+    bool wakeup = false;
+
     OS_UTILS_LockMutex(&mqtt_access_mutex);
 
     if (is_mqtt_mtp_thread_exited)
@@ -2455,16 +2504,22 @@ void MQTT_ActivateScheduledActions(void)
         if (client->schedule_reconnect == kScheduledAction_Signalled)
         {
             client->schedule_reconnect = kScheduledAction_Activated;
+            wakeup = true;
+        }
 
-            OS_UTILS_UnlockMutex(&mqtt_access_mutex);
-            MTP_EXEC_MqttWakeup();
-            return;
+        if (client->schedule_close == kScheduledAction_Signalled)
+        {
+            client->schedule_close = kScheduledAction_Activated;
+            wakeup = true;
         }
     }
 
     OS_UTILS_UnlockMutex(&mqtt_access_mutex);
 
-    return;
+    if (wakeup)
+    {
+        MTP_EXEC_MqttWakeup();
+    }
 }
 
 mtp_status_t MQTT_GetMtpStatus(int instance)
@@ -2724,6 +2779,59 @@ int MQTT_ScheduleResubscription(int instance, mqtt_subscription_t *subscription)
     }
 
     return err;
+}
+
+/*********************************************************************//**
+**
+** MQTT_GetAgentResponseTopicDiscovered
+**
+** Reads the value of the CONNACK Response Information property supplied by a MQTT 5.0 broker
+** If this is not available (for example not MQTT v5.0 or CONNACK not received yet) then an empty string is returned
+**
+** \param   instance - instance in Device.MQTT.Client.{i}
+** \param   buf - pointer to buffer into which to return the value of the parameter (as a textual string)
+** \param   len - length of buffer in which to return the value of the parameter
+**
+** \return  Always USP_ERR_OK - an empty string is returned if the value cannot be determined
+**
+**************************************************************************/
+int MQTT_GetAgentResponseTopicDiscovered(int instance, char *buf, int len)
+{
+    mqtt_client_t *client;
+
+    OS_UTILS_LockMutex(&mqtt_access_mutex);
+
+    // Set default return value - an empty string
+    *buf = '\0';
+
+    // Exit if no client exists with the specified instance number
+    client = FindMqttClientByInstance(instance);
+    if (client == NULL)
+    {
+        goto exit;
+    }
+
+    // Exit if client is not currently connected
+    if (client->state != kMqttState_Running)
+    {
+        goto exit;
+    }
+
+    // Exit if client is not using MQTT v5 (earlier versions of MQTT do not allow for a response information property in the CONNACK)
+    if (client->conn_params.version != kMqttProtocol_5_0)
+    {
+        goto exit;
+    }
+
+    // Copy the agent's discovered response topic into the return buffer
+    if (client->agent_topic_from_connack != NULL)
+    {
+        USP_STRNCPY(buf, client->agent_topic_from_connack, len);
+    }
+
+exit:
+    OS_UTILS_UnlockMutex(&mqtt_access_mutex);
+    return USP_ERR_OK;
 }
 
 #endif